+++ test/customEvent.test.js
@@ -28,11 +28,56 @@ describe('customEvent method', () => {
         user: '',
         warehouse: 'Belgrano',
         rol: 'picker',
+        language: 'EN-US',
       };
 
       const event = await customEvent('event_name', validParams);
 
       expect(event).toBe(true);
+    });
+  });
+
+  describe('can receive a custom string array that represent required params for some custom event', () => {
+    describe('this array is validate and', () => {
+      it('is invalid argument, should will be replaced by an empty array and the action complete normally', async () => {
+        const validParams = {
+          actionName: 'buttonPress',
+          client: 'janis',
+          userEmail: 'janis@janis.im',
+          userId: 'user2022Janis',
+          appVersion: '1.20.0',
+          user: '',
+          warehouse: 'Belgrano',
+          rol: 'picker',
+          language: 'EN-US',
+        };
+
+        const event = await customEvent('event_name', validParams, {});
+
+        expect(event).toBe(true);
+      });
+
+      it('is valid argument, filter invalid array elements and pass remaining for validation', async () => {
+        const validParams = {
+          actionName: 'buttonPress',
+          client: 'janis',
+          userEmail: 'janis@janis.im',
+          userId: 'user2022Janis',
+          appVersion: '1.20.0',
+          user: '',
+          warehouse: 'Belgrano',
+          rol: 'picker',
+          language: 'EN-US',
+        };
+
+        const event = await customEvent('event_name', validParams, [
+          'rol',
+          3,
+          'actionName',
+        ]);
+
+        expect(event).toBe(true);
+      });
     });
   });
 });